refactor(extensions): move window.superset augmentation to its own module
The `window.superset` global type augmentation lived inline in
ExtensionsStartup, but ExtensionsLoader also reads/writes `window.superset`.
When TypeScript compiles a file set that includes ExtensionsLoader without
ExtensionsStartup (e.g. the pre-commit hook's narrowed batch), the augmentation
is out of scope and `window.superset` fails to type-check.

Extract the augmentation into a dedicated `supersetGlobal` module that both
ExtensionsStartup and ExtensionsLoader import for its side effect, so the type
travels with the import graph and resolves regardless of compilation batching.
